+++ src/common/factory.ts
@@ -14,11 +14,34 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-import { TurboUnauthenticatedConfiguration } from '../types.js';
+import { ArweaveSigner, EthereumSigner, HexSolanaSigner } from 'arbundles';
+
+import { TurboNodeSigner } from '../node/signer.js';
+import {
+  CreditableTokenType,
+  TurboAuthenticatedConfiguration,
+  TurboSigner,
+  TurboUnauthenticatedConfiguration,
+  TurboWallet,
+  isJWK,
+} from '../types.js';
+import { TurboWebArweaveSigner } from '../web/signer.js';
+import { JWKInterface } from './jwk.js';
 import { TurboWinstonLogger } from './logger.js';
-import { TurboUnauthenticatedPaymentService } from './payment.js';
-import { TurboUnauthenticatedClient } from './turbo.js';
-import { TurboUnauthenticatedUploadService } from './upload.js';
+import {
+  TurboAuthenticatedPaymentService,
+  TurboUnauthenticatedPaymentService,
+} from './payment.js';
+import { TurboDataItemAbstractSigner } from './signer.js';
+import { defaultTokenMap } from './token/index.js';
+import {
+  TurboAuthenticatedClient,
+  TurboUnauthenticatedClient,
+} from './turbo.js';
+import {
+  TurboAuthenticatedUploadService,
+  TurboUnauthenticatedUploadService,
+} from './upload.js';
 
 export class TurboBaseFactory {
   protected static logger = new TurboWinstonLogger();
@@ -44,6 +67,97 @@ export class TurboBaseFactory {
       logger: this.logger,
     });
     return new TurboUnauthenticatedClient({
+      uploadService,
+      paymentService,
+    });
+  }
+
+  protected static getSigner(
+    providedSigner: TurboSigner | undefined,
+    providedPrivateKey: TurboWallet | undefined,
+    token: CreditableTokenType,
+  ): TurboDataItemAbstractSigner {
+    let signer: TurboSigner;
+
+    if (providedSigner !== undefined) {
+      signer = providedSigner;
+    } else if (providedPrivateKey !== undefined) {
+      if (token === 'solana') {
+        signer = new HexSolanaSigner(providedPrivateKey);
+        // TODO: else if (token === 'ethereum') {signer = new EthereumSigner(providedPrivateKey);}
+      } else {
+        if (!isJWK(providedPrivateKey)) {
+          throw new Error('A JWK must be provided for ArweaveSigner.');
+        }
+        signer = new ArweaveSigner(providedPrivateKey as JWKInterface);
+      }
+    } else {
+      throw new Error('A privateKey or signer must be provided.');
+    }
+
+    if (typeof window !== 'undefined') {
+      return new TurboWebArweaveSigner({
+        signer,
+        logger: this.logger,
+      });
+    }
+    return new TurboNodeSigner({
+      signer,
+      logger: this.logger,
+    });
+  }
+
+  static authenticated({
+    privateKey,
+    signer: providedSigner,
+    paymentServiceConfig = {},
+    uploadServiceConfig = {},
+    token,
+    tokenMap,
+    gatewayUrl,
+    tokenTools,
+  }: TurboAuthenticatedConfiguration) {
+    if (!token) {
+      if (providedSigner) {
+        // Derive token from signer if not provided
+        if (providedSigner instanceof EthereumSigner) {
+          token = 'ethereum';
+        } else if (providedSigner instanceof HexSolanaSigner) {
+          token = 'solana';
+        } else {
+          token = 'arweave';
+        }
+      } else {
+        token = 'arweave';
+      }
+    }
+
+    const turboSigner = this.getSigner(providedSigner, privateKey, token);
+
+    if (!tokenTools) {
+      if (tokenMap && token === 'arweave') {
+        tokenTools = tokenMap.arweave;
+      }
+      tokenTools = defaultTokenMap[token]?.({
+        gatewayUrl,
+        logger: this.logger,
+      });
+    }
+
+    const paymentService = new TurboAuthenticatedPaymentService({
+      ...paymentServiceConfig,
+      signer: turboSigner,
+      logger: this.logger,
+      token,
+      tokenTools,
+    });
+    const uploadService = new TurboAuthenticatedUploadService({
+      ...uploadServiceConfig,
+      signer: turboSigner,
+      logger: this.logger,
+      token,
+    });
+    return new TurboAuthenticatedClient({
       uploadService,
       paymentService,
     });