+++ src/types.ts
@@ -189,7 +189,13 @@ export type TurboPostBalanceResponse =
       message: string;
     };
 
-export type TurboWallet = JWKInterface; // TODO: add other wallet types
+type Base58String = string;
+export type SolSecretKey = Base58String;
+export type TurboWallet = JWKInterface | SolSecretKey; // TODO: add other wallet types
+
+export const isJWK = (wallet: TurboWallet): wallet is JWKInterface =>
+  (wallet as JWKInterface).kty !== undefined;
+
 export type TurboSignedRequestHeaders = {
   'x-public-key': string;
   'x-nonce': string;
@@ -253,6 +259,8 @@ export type TurboAuthenticatedConfiguration =
   TurboUnauthenticatedConfiguration & {
     privateKey?: TurboWallet;
     signer?: TurboSigner;
+    /** @deprecated -- This parameter was added in release v1.5 for injecting an arweave TokenTool. Instead, the SDK now accepts `tokenTools` and/or `gatewayUrl` directly in the Factory constructor. This type will be removed in a v2 release */
+    tokenMap?: TokenMap;
     tokenTools?: TokenTools;
     token?: CreditableTokenType;
     gatewayUrl?: string;
@@ -424,3 +432,14 @@ export interface TokenTools {
   }>;
   pollForTxBeingAvailable: (p: { txId: string }) => Promise<void>;
 }
+
+export type TokenConfig = {
+  gatewayUrl?: string;
+  logger?: TurboLogger;
+  pollingOptions?: TokenPollingOptions;
+};
+
+/** @deprecated -- This type was provided as a parameter in release v1.5 for injecting an arweave TokenTool. Instead, the SDK now accepts `tokenTools` and/or `gatewayUrl`  directly in the Factory constructor. This type will be removed in a v2 release  */
+export type TokenMap = { arweave: TokenTools };
+
+export type TokenFactory = Record<string, (config: TokenConfig) => TokenTools>;